refactor: Move PrinterRenderer to services module and update imports
